@@ -1,4 +1,11 @@
-__version__ = "1.5.0.post20230715"
+__version__ = "2.0.0.post20230716"
 
-from .api import Api, Base, Table  # noqa
-from .api.retrying import retry_strategy  # noqa
+from .api import Api, Base, Table
+from .api.retrying import retry_strategy
+
+__all__ = [
+    "Api",
+    "Base",
+    "Table",
+    "retry_strategy",
+]
