@@ -1,1 +1,7 @@
-from .model import Model  # noqa
+from . import fields
+from .model import Model
+
+__all__ = [
+    "Model",
+    "fields",
+]
