@@ -1,76 +1,23 @@
-"""
-The :class:`Model` class allows you create an orm-style class for your
-Airtable tables.
+from functools import lru_cache
+from typing import Any, Dict, Iterable, List, Optional
 
->>> from pyairtable.orm import Model, fields
->>> class Contact(Model):
-...     first_name = fields.TextField("First Name")
-...     last_name = fields.TextField("Last Name")
-...     email = fields.EmailField("Email")
-...     is_registered = fields.CheckboxField("Registered")
-...     company = fields.LinkField("Company", Company, lazy=False)
-...
-...     class Meta:
-...         base_id = "appaPqizdsNHDvlEm"
-...         table_name = "Contact"
-...         api_key = "keyapikey"
+from typing_extensions import Self as SelfType
 
-
-Once you have a class, you can create new objects to represent your
-Airtable records. Call :meth:`~pyairtable.orm.model.Model.save` to create a new record.
-
->>> contact = Contact(
-...     first_name="Mike",
-...     last_name="McDonalds",
-...     email="mike@mcd.com",
-...     is_registered=False
-... )
-...
->>> assert contact.id is None
->>> contact.exists()
-False
->>> assert contact.save()
->>> contact.exists()
-True
->>> contact.id
-rec123asa23
+from pyairtable.api.api import Api
+from pyairtable.api.base import Base
+from pyairtable.api.table import Table
+from pyairtable.api.types import (
+    FieldName,
+    RecordDict,
+    RecordId,
+    UpdateRecordDict,
+    WritableFields,
+)
+from pyairtable.formulas import OR, STR_VALUE
+from pyairtable.orm.fields import AnyField, Field
 
 
-You can read and modify attributes. If record already exists,
-:meth:`~pyairtable.orm.model.Model.save` will update the record:
-
->>> assert contact.is_registered is False
->>> contact.save()
->>> assert contact.is_registered is True
->>> contact.to_record()
-{
-    "id": recS6qSLw0OCA6Xul",
-    "createdTime": "2021-07-14T06:42:37.000Z",
-    "fields": {
-        "First Name": "Mike",
-        "Last Name": "McDonalds",
-        "Email": "mike@mcd.com",
-        "Resgistered": True
-    }
-}
-
-Finally, you can use :meth:`~pyairtable.orm.model.Model.delete` to delete the record:
-
->>> contact.delete()
-True
-
-"""
-import abc
-from typing import List, Type, TypeVar
-
-from pyairtable import Table
-
-from .fields import Field
-
-T = TypeVar("T", bound="Model")
-
-
-class Model(metaclass=abc.ABCMeta):
+class Model:
     """
     This class allows you to create an orm-style class for your Airtable tables.
 
@@ -94,16 +41,20 @@ class Model(metaclass=abc.ABCMeta):
 
     id: str = ""
     created_time: str = ""
-    _table: Table
-    _fields: dict = {}
-    _linked_cache: dict = {}
+    _deleted: bool = False
+    _fields: Dict[FieldName, Any]
 
-    def __init_subclass__(cls, **kwargs):
+    def __init_subclass__(cls, **kwargs: Any):
         cls._validate_class()
         super().__init_subclass__(**kwargs)
 
+    def __repr__(self) -> str:
+        if not self.id:
+            return f"<unsaved {self.__class__.__name__}>"
+        return f"<{self.__class__.__name__} id={self.id!r}>"
+
     @classmethod
-    def _attribute_descriptor_map(cls):
+    def _attribute_descriptor_map(cls) -> Dict[str, AnyField]:
         """
         Returns a dictionary mapping the model's attribute names to the field's
 
@@ -120,9 +71,9 @@ class Model(metaclass=abc.ABCMeta):
         return {k: v for k, v in cls.__dict__.items() if isinstance(v, Field)}
 
     @classmethod
-    def _field_name_descriptor_map(cls):
+    def _field_name_descriptor_map(cls) -> Dict[FieldName, AnyField]:
         """
-        Returns a dictionary that maps Fields 'Names' to descriptor fields
+        Returns a dictionary that maps field names to descriptor instances.
 
         >>> class Test(Model):
         ...     first_name = TextField("First Name")
@@ -137,9 +88,9 @@ class Model(metaclass=abc.ABCMeta):
         return {f.field_name: f for f in cls._attribute_descriptor_map().values()}
 
     @classmethod
-    def _field_name_attribute_map(cls):
+    def _field_name_attribute_map(cls) -> Dict[FieldName, str]:
         """
-        Returns a dictionary that maps Fields 'Names' to the model attribute name:
+        Returns a dictionary that maps field names to attribute names.
 
         >>> class Test(Model):
         ...     first_name = TextField("First Name")
@@ -153,29 +104,36 @@ class Model(metaclass=abc.ABCMeta):
         """
         return {v.field_name: k for k, v in cls._attribute_descriptor_map().items()}
 
-    def __init__(self, **fields):
-        # To Store Fields
+    def __init__(self, **fields: Any):
+        if "id" in fields:
+            self.id = fields.pop("id")
+
+        # Field values in internal (not API) representation
         self._fields = {}
 
-        # Set descriptors
+        # Call __set__ on each field to set field values
         for key, value in fields.items():
             if key not in self._attribute_descriptor_map():
-                msg = "invalid kwarg '{}'".format(key)
-                raise ValueError(msg)
+                raise AttributeError(key)
             setattr(self, key, value)
 
-        self.typecast = getattr(self.Meta, "typecast", True)
-        self.raise_key_error = getattr(self.Meta, "raise_key_error", True)
+    @classmethod
+    def _get_meta(cls, name: str, default: Any = None, required: bool = False) -> Any:
+        if not hasattr(cls, "Meta"):
+            raise AttributeError(f"{cls.__name__}.Meta must be defined")
+        if required and not hasattr(cls.Meta, name):
+            raise ValueError(f"{cls.__name__}.Meta.{name} must be defined")
+        value = getattr(cls.Meta, name, default)
+        if required and value is None:
+            raise ValueError(f"{cls.__name__}.Meta.{name} cannot be None")
+        return value
 
     @classmethod
-    def _validate_class(cls):
+    def _validate_class(cls) -> None:
         # Verify required Meta attributes were set
-        if not getattr(cls.Meta, "base_id", None):
-            raise ValueError("Meta.base_id must be defined in model")
-        if not getattr(cls.Meta, "table_name", None):
-            raise ValueError("Meta.table_name must be defined in model")
-        if not getattr(cls.Meta, "api_key", None):
-            raise ValueError("Meta.api_key must be defined in model")
+        assert cls._get_meta("api_key", required=True)
+        assert cls._get_meta("base_id", required=True)
+        assert cls._get_meta("table_name", required=True)
 
         model_attributes = [a for a in cls.__dict__.keys() if not a.startswith("__")]
         overridden = set(model_attributes).intersection(Model.__dict__.keys())
@@ -187,16 +145,24 @@ class Model(metaclass=abc.ABCMeta):
             )
 
     @classmethod
+    @lru_cache
+    def get_api(cls) -> Api:
+        return Api(
+            api_key=cls._get_meta("api_key"),
+            timeout=cls._get_meta("timeout"),
+        )
+
+    @classmethod
+    def get_base(cls) -> Base:
+        return cls.get_api().base(cls._get_meta("base_id"))
+
+    @classmethod
     def get_table(cls) -> Table:
-        """Return Airtable :class:`~pyairtable.api.Table` class instance"""
-        if not hasattr(cls, "_table"):
-            cls._table = Table(
-                cls.Meta.api_key,  # type: ignore
-                cls.Meta.base_id,  # type: ignore
-                cls.Meta.table_name,  # type: ignore
-                timeout=getattr(cls.Meta, "timeout", None),  # type: ignore
-            )
-        return cls._table
+        return cls.get_base().table(cls._get_meta("table_name"))
+
+    @classmethod
+    def _typecast(cls) -> bool:
+        return bool(cls._get_meta("typecast", default=True))
 
     def exists(self) -> bool:
         """Returns boolean indicating if instance exists (has 'id' attribute)"""
@@ -205,19 +171,20 @@ class Model(metaclass=abc.ABCMeta):
     def save(self) -> bool:
         """
         Saves or updates a model.
-        If instance has no 'id', it will be created, otherwise updatedself.
+        If instance has no 'id', it will be created, otherwise updated.
 
-        Returns `True` if was created and `False` if it was updated
+        Returns ``True`` if was created and `False` if it was updated
         """
+        if self._deleted:
+            raise RuntimeError(f"{self.id} was deleted")
         table = self.get_table()
-        record = self.to_record()
-        fields = record["fields"]
+        fields = self.to_record(only_writable=True)["fields"]
 
         if not self.id:
-            record = table.create(fields, typecast=self.typecast)
+            record = table.create(fields, typecast=self._typecast())
             did_create = True
         else:
-            record = table.update(self.id, fields, typecast=self.typecast)
+            record = table.update(self.id, fields, typecast=self._typecast())
             did_create = False
 
         self.id = record["id"]
@@ -225,74 +192,88 @@ class Model(metaclass=abc.ABCMeta):
         return did_create
 
     def delete(self) -> bool:
-        """Deleted record. Must have 'id' field"""
+        """Deletes record. Must have 'id' field"""
         if not self.id:
             raise ValueError("cannot be deleted because it does not have id")
         table = self.get_table()
         result = table.delete(self.id)
-        # Is it even possible go get "deleted" False?
-        return result["deleted"]
+        self._deleted = True
+        # Is it even possible to get "deleted" False?
+        return bool(result["deleted"])
 
     @classmethod
-    def all(cls, **kwargs) -> List[T]:
-        """Returns all records for this model. See :meth:`~pyairtable.api.Api.all`"""
+    def all(cls, **kwargs: Any) -> List[SelfType]:
+        """
+        Returns all records for this model. For all supported
+        keyword arguments, see :meth:`Table.all <pyairtable.Table.all>`.
+        """
         table = cls.get_table()
-        return table.all(**kwargs)
+        return [cls.from_record(record) for record in table.all(**kwargs)]
 
     @classmethod
-    def first(cls, **kwargs) -> List[T]:
-        """Returns first record for this model. See :meth:`~pyairtable.api.Api.first`"""
+    def first(cls, **kwargs: Any) -> Optional[SelfType]:
+        """
+        Returns the first record for this model. For all supported
+        keyword arguments, see :meth:`Table.first <pyairtable.Table.first>`.
+        """
         table = cls.get_table()
-        return table.first(**kwargs)
+        if record := table.first(**kwargs):
+            return cls.from_record(record)
+        return None
 
-    def to_record(self) -> dict:
+    def to_record(self, only_writable: bool = False) -> RecordDict:
         """
         Returns a dictionary object as an Airtable record.
         This method converts internal field values into values expected by Airtable.
-        e.g. a ``datetime`` value from :class:``DateTimeField`` is converted into an
-        ISO 8601 string
+        For example, a ``datetime`` value from :class:`~pyairtable.orm.fields.DatetimeField`
+        is converted into an ISO 8601 string.
+
+        Args:
+            only_writable: If ``True``, the result will exclude any
+                values which are associated with readonly fields.
         """
         map_ = self._field_name_descriptor_map()
-        fields = {k: map_[k].to_record_value(v) for k, v in self._fields.items()}
+        fields = {
+            field: None if value is None else map_[field].to_record_value(value)
+            for field, value in self._fields.items()
+            if not (map_[field].readonly and only_writable)
+        }
         return {"id": self.id, "createdTime": self.created_time, "fields": fields}
 
     @classmethod
-    def from_record(cls: Type[T], record: dict) -> T:
-        """Create instance from record dictionary"""
-        name_attr_map = cls._field_name_attribute_map()
+    def from_record(cls, record: RecordDict) -> SelfType:
+        """
+        Create an instance from a record dict.
+        """
         name_field_map = cls._field_name_descriptor_map()
-        raise_key_error = getattr(cls.Meta, 'raise_key_error', True)
 
-        if raise_key_error:
-            try:
-                # Convert Column Names into model field names
-                # Use field's to_internal to cast into model fields
-                kwargs = {
-                    name_attr_map[k]: name_field_map[k].to_internal_value(v)
-                    for k, v in record["fields"].items()
-                    # Silently proceed if Airtable returns fields we don't recognize
-                    if k in name_attr_map
-                }
-            except KeyError as exc:
-                raise ValueError("Invalid Field Name: {} for model {}".format(exc, cls))
-        else:
-            kwargs = {
-                name_attr_map[k]: name_field_map[k].to_internal_value(v)
-                for k, v in record["fields"].items() if k in name_attr_map
-            }
-
-        instance = cls(**kwargs)
+        # Convert Column Names into model field names
+        field_values = {
+            # Use field's to_internal_value to cast into model fields
+            field: name_field_map[field].to_internal_value(value)
+            for (field, value) in record["fields"].items()
+            # Silently proceed if Airtable returns fields we don't recognize
+            if field in name_field_map
+        }
+        # Since instance(**field_values) will perform validation and fail on
+        # any readonly fields, instead we directly set instance._fields.
+        instance = cls(id=record["id"])
+        instance._fields = field_values
         instance.created_time = record["createdTime"]
-        instance.id = record["id"]
         return instance
 
     @classmethod
-    def from_id(cls: Type[T], record_id: str, fetch=True) -> T:
+    def from_id(
+        cls,
+        record_id: RecordId,
+        fetch: bool = True,
+    ) -> SelfType:
         """
-        Create an instance from a `record_id`
+        Create an instance from a record ID.
 
         Args:
             record_id: |arg_record_id|
+
             fetch:
 
         Keyword Args:
@@ -302,52 +283,96 @@ class Model(metaclass=abc.ABCMeta):
 
         Returns:
             (``Model``): Instance of model
-        """
-        if fetch:
-            table = cls.get_table()
-            record = table.get(record_id)
-            return cls.from_record(record)
-        else:
-            instance = cls()
-            instance.id = record_id
-            return instance
 
-    def fetch(self):
-        """Fetches field and resets instance field values from the Airtable record"""
+            fetch: If ``True``, record will be fetched and field values will be
+                updated. If ``False``, a new instance is created with the provided ID,
+                but field values are unset.
+        """
+        instance = cls(id=record_id)
+        if fetch:
+            instance.fetch()
+        return instance
+
+    def fetch(self) -> None:
+        """
+        Fetches field values from the API and resets instance field values.
+        """
         if not self.id:
             raise ValueError("cannot be fetched because instance does not have an id")
 
-        updated = self.from_id(self.id, fetch=True)
-        self._fields = updated._fields
-        self.created_time = updated.created_time
+        record = self.get_table().get(self.id)
+        unused = self.from_record(record)
+        self._fields = unused._fields
+        self.created_time = unused.created_time
 
-    def __repr__(self):
-        return "<Model={} {}>".format(self.__class__.__name__, hex(id(self)))
+    @classmethod
+    def from_ids(
+        cls,
+        record_ids: Iterable[RecordId],
+        fetch: bool = True,
+    ) -> List[SelfType]:
+        """
+        Create a list of instances from record IDs. If any record IDs returned
+        are invalid this will raise a KeyError, but only *after* retrieving all
+        other valid records from the API.
 
-    # TODO - see metadata.py
-    # def verify_schema(cls) -> Tuple[bool, dict]:
-    #     """verify local airtable models"""
+        Args:
+            record_ids: |arg_record_id|
+            fetch: If ``True``, records will be fetched and field values will be
+                updated. If ``False``, new instances are created with the provided IDs,
+                but field values are unset.
+        """
+        record_ids = list(record_ids)
+        if not fetch:
+            return [cls.from_id(record_id, fetch=False) for record_id in record_ids]
+        formula = OR(
+            *[f"RECORD_ID()={STR_VALUE(record_id)}" for record_id in record_ids]
+        )
+        records = [
+            cls.from_record(record) for record in cls.get_table().all(formula=formula)
+        ]
+        records_by_id = {record.id: record for record in records}
+        # Ensure we return records in the same order, and raise KeyError if any are missing
+        return [records_by_id[record_id] for record_id in record_ids]
 
-    #     base_list = cls.get_base_list()
-    #     base_id_exists = cls.base_id in [b["id"] for b in base_list["bases"]]
+    @classmethod
+    def batch_save(cls, models: List[SelfType]) -> None:
+        """
+        Saves a list of model instances to the Airtable API with as few
+        network requests as possible. Can accept a mixture of new records
+        (which have not been saved yet) and existing records that have IDs.
+        """
+        if not all(isinstance(model, cls) for model in models):
+            raise TypeError(set(type(model) for model in models))
 
-    #     if base_id_exists:
-    #         base_schema = cls.get_base_schema()
-    #         table_schema: dict = next(
-    #             (t for t in base_schema["tables"] if t["name"] == cls.table_name), {}
-    #         )
-    #         table_exists = bool(table_schema)
-    #     else:
-    #         table_exists = False
+        create_models = [model for model in models if not model.id]
+        update_models = [model for model in models if model.id]
+        create_records: List[WritableFields] = [
+            record["fields"]
+            for model in create_models
+            if (record := model.to_record(only_writable=True))
+        ]
+        update_records: List[UpdateRecordDict] = [
+            {"id": record["id"], "fields": record["fields"]}
+            for model in update_models
+            if (record := model.to_record(only_writable=True))
+        ]
 
-    #     if table_exists:
-    #         airtable_field_names = [f["name"] for f in table_schema["fields"]]
-    #         # Fields.NAME = "|NAME|", Fields.ID = "|ID|") -> ["|NAME|", "|ID|"]
-    #         local_field_names = [v for k, v in vars(cls.Fields).items() if k.isupper()]
-    #         fields = {n: n in airtable_field_names for n in local_field_names}
-    #     else:
-    #         fields = {}
+        table = cls.get_table()
+        table.batch_update(update_records, typecast=cls._typecast())
+        created_records = table.batch_create(create_records, typecast=cls._typecast())
+        for model, created_record in zip(create_models, created_records):
+            model.id = created_record["id"]
+            model.created_time = created_record["createdTime"]
 
-    #     in_sync = base_id_exists and table_exists and all(fields.values())
-    #     details = {"base": base_id_exists, "table": table_exists, "fields": fields}
-    #     return (in_sync, details)
+    @classmethod
+    def batch_delete(cls, models: List[SelfType]) -> None:
+        """
+        Deletes a list of model instances from Airtable. Raises ``ValueError`` if
+        given a model which has never been saved to Airtable.
+        """
+        if not all(model.id for model in models):
+            raise ValueError("cannot delete an unsaved model")
+        if not all(isinstance(model, cls) for model in models):
+            raise TypeError(set(type(model) for model in models))
+        cls.get_table().batch_delete([model.id for model in models])
