@@ -1,5 +1,9 @@
 from datetime import date, datetime
-from typing import Union
+from typing import Iterator, Sequence, TypeVar, Union
+
+from pyairtable.api.types import CreateAttachmentDict
+
+T = TypeVar("T")
 
 
 def datetime_to_iso_str(value: datetime) -> str:
@@ -15,22 +19,20 @@ def datetime_to_iso_str(value: datetime) -> str:
 
 def datetime_from_iso_str(value: str) -> datetime:
     """
-    Converts ISO 8601 datetime string into a ``datetime`` object.
-    Expected format is "2014-09-05T07:00:00.000Z"
+    Convert an ISO 8601 datetime string into a ``datetime`` object.
 
     Args:
-        value: datetime string e.g. "2014-09-05T07:00:00.000Z"
+        value: datetime string, e.g. "2014-09-05T07:00:00.000Z"
     """
     return datetime.strptime(value, "%Y-%m-%dT%H:%M:%S.%fZ")
 
 
 def date_to_iso_str(value: Union[date, datetime]) -> str:
     """
-    Converts ``date`` or ``datetime`` object into Airtable compatible ISO 8601 string
-    e.g. "2014-09-05"
+    Converts a ``date`` or ``datetime`` into an Airtable-compatible ISO 8601 string
 
     Args:
-        value: date or datetime object
+        value: date or datetime object, e.g. "2014-09-05"
     """
     return value.strftime("%Y-%m-%d")
 
@@ -38,24 +40,26 @@ def date_to_iso_str(value: Union[date, datetime]) -> str:
 def date_from_iso_str(value: str) -> date:
     """
     Converts ISO 8601 date string into a ``date`` object.
-    Expected format is  "2014-09-05"
 
     Args:
-        value: date string e.g. "2014-09-05"
+        value: date string, e.g. "2014-09-05"
     """
     return datetime.strptime(value, "%Y-%m-%d").date()
 
 
-def attachment(url: str, filename="") -> dict:
+def attachment(url: str, filename: str = "") -> CreateAttachmentDict:
     """
-    Returns a dictionary using the expected dicitonary format for attachments.
+    Returns a dictionary using the expected dictionary format for creating attachments.
 
     When creating an attachment, ``url`` is required, and ``filename`` is optional.
     Airtable will download the file at the given url and keep its own copy of it.
     All other attachment object properties will be generated server-side soon afterward.
 
     Note:
-        Attachment field values muest be **an array of objects**.
+        Attachment field values **must** be an array of
+        :class:`~pyairtable.api.types.AttachmentDict` or
+        :class:`~pyairtable.api.types.CreateAttachmentDict`;
+        it is not valid to pass a single item to the API.
 
     Usage:
         >>> table = Table(...)
@@ -78,3 +82,15 @@ def attachment(url: str, filename="") -> dict:
 
     """
     return {"url": url} if not filename else {"url": url, "filename": filename}
+
+
+def chunked(iterable: Sequence[T], chunk_size: int) -> Iterator[Sequence[T]]:
+    """
+    Break a sequence into chunks.
+
+    Args:
+        iterable: Any sequence.
+        chunk_size: Maximum items to yield per chunk.
+    """
+    for i in range(0, len(iterable), chunk_size):
+        yield iterable[i : i + chunk_size]
