@@ -32,23 +32,18 @@ def constants():
 
 
 @pytest.fixture()
-def api(constants):
+def api(constants) -> Api:
     return Api(constants["API_KEY"])
 
 
 @pytest.fixture()
-def api_with_endpoint_url(constants):
-    return Api(constants["API_KEY"], endpoint_url="https://api.example.com")
+def base(api: Api, constants) -> Base:
+    return api.base(constants["BASE_ID"])
 
 
 @pytest.fixture()
-def base(constants):
-    return Base(constants["API_KEY"], constants["BASE_ID"])
-
-
-@pytest.fixture()
-def table(constants):
-    return Table(constants["API_KEY"], constants["BASE_ID"], constants["TABLE_NAME"])
+def table(base: Base, constants) -> Table:
+    return base.table(constants["TABLE_NAME"])
 
 
 @pytest.fixture
