@@ -2,14 +2,7 @@ import os
 
 import pytest
 
-from pyairtable import Base, Table
-
-BASE_ID = "appaPqizdsNHDvlEm"
-
-
-@pytest.fixture
-def base_name():
-    return "Test Wrapper"
+from pyairtable import Api, Base, Table
 
 
 @pytest.fixture
@@ -33,20 +26,54 @@ def cols():
 
 
 @pytest.fixture
-def base():
-    base_id = BASE_ID
-    base = Base(os.environ["AIRTABLE_API_KEY"], base_id)
-    yield base
-    table_name = "TEST_TABLE"
-    records = base.all(table_name)
-    base.batch_delete(table_name, [r["id"] for r in records])
+def api_key():
+    try:
+        return os.environ["AIRTABLE_API_KEY"]
+    except KeyError:
+        pytest.skip("integration test requires AIRTABLE_API_KEY env variable")
 
 
 @pytest.fixture
-def table():
-    base_id = BASE_ID
-    table_name = "TEST_TABLE"
-    table = Table(os.environ["AIRTABLE_API_KEY"], base_id, table_name)
+def api(api_key) -> Api:
+    return Api(api_key)
+
+
+@pytest.fixture
+def base_id():
+    return "appaPqizdsNHDvlEm"
+
+
+@pytest.fixture
+def base(api, base_id) -> Base:
+    return api.base(base_id)
+
+
+@pytest.fixture
+def base_name():
+    return "Test Wrapper"
+
+
+@pytest.fixture
+def table_name():
+    return "TEST_TABLE"
+
+
+@pytest.fixture
+def table(base: Base, table_name) -> Table:
+    table = base.table(table_name)
     yield table
     records = table.all()
     table.batch_delete([r["id"] for r in records])
+
+
+@pytest.fixture
+def make_meta(api_key, base_id):
+    def _make_meta(table_name):
+        dct = {
+            "api_key": api_key,
+            "base_id": base_id,
+            "table_name": table_name,
+        }
+        return type("Meta", (), dct)
+
+    return _make_meta
