@@ -1,50 +1,39 @@
-from unittest import mock
+import pytest
 
 from pyairtable import Base, Table
-from pyairtable.api.abstract import ApiAbstract
+
+
+def test_constructor(api):
+    base = Base(api, "base_id")
+    assert base.api == api
+    assert base.id == "base_id"
+
+
+def test_deprecated_constructor():
+    with pytest.warns(DeprecationWarning):
+        base = Base("api_key", "base_id")
+
+    assert base.api.api_key == "api_key"
+    assert base.id == "base_id"
+
+
+def test_invalid_constructor():
+    """
+    Test that we get a TypeError if passing invalid kwargs to Base.
+    """
+    with pytest.raises(TypeError):
+        Base(api_key="api_key", base_id="base_id")
+    with pytest.raises(TypeError):
+        Base("api_key", "base_id", timeout=(1, 1))
 
 
 def test_repr(base):
     assert "Base" in base.__repr__()
 
 
-def test_record_url(base: Base):
-    rv = base.get_record_url("tablename", "rec")
-    assert rv == f"https://api.airtable.com/v0/{base.base_id}/tablename/rec"
-
-
 def test_get_table(base: Base):
-    rv = base.get_table("tablename")
+    rv = base.table("tablename")
     assert isinstance(rv, Table)
-    assert rv.base_id == base.base_id
-    assert rv.table_name == "tablename"
-    assert rv.table_url == f"https://api.airtable.com/v0/{base.base_id}/tablename"
-
-
-@mock.patch.object(ApiAbstract, "_get_record")
-def test_get(m, base: Base, mock_response_single):
-    m.return_value = mock_response_single
-    rv = base.get("tablename", "rec")
-    assert rv == mock_response_single
-
-
-@mock.patch.object(ApiAbstract, "_first")
-def test_first(m, base: Base, mock_response_single):
-    m.return_value = mock_response_single
-    rv = base.first("tablename")
-    assert rv == mock_response_single
-
-
-@mock.patch.object(ApiAbstract, "_all")
-def test_all(m, base: Base, mock_response_list):
-    m.return_value = mock_response_list
-    rv = base.all("tablename")
-    assert rv == mock_response_list
-
-
-@mock.patch.object(ApiAbstract, "_update")
-def test_update(m, base: Base, mock_response_single):
-    m.return_value = mock_response_single
-    rv = base.update("tablename", "rec", {"test": "test"})
-    assert rv == mock_response_single
-    assert {"test": "test"} in m.call_args[0][:4]
+    assert rv.base == base
+    assert rv.name == "tablename"
+    assert rv.url == f"https://api.airtable.com/v0/{base.id}/tablename"
