@@ -8,7 +8,6 @@ from pyairtable.api.params import (
     field_names_to_sorting_dict,
     options_to_json_and_params,
     options_to_params,
-    to_params_dict,
 )
 
 
@@ -31,7 +30,7 @@ def test_params_integration(table, mock_records, mock_response_iterator):
             "&returnFieldsByFieldId=1"
             ""
         )
-        mock_url = "{0}?{1}".format(table.table_url, url_params)
+        mock_url = "{0}?{1}".format(table.url, url_params)
         m.get(mock_url, status_code=200, json=mock_response_iterator)
         response = table.all(**params)
     for n, resp in enumerate(response):
@@ -115,13 +114,6 @@ def test_params_integration(table, mock_records, mock_response_iterator):
 def test_convert_options_to_params(option, value, url_params):
     """Ensure kwargs received build a proper params"""
     processed_params = options_to_params({option: value})
-    request = requests.Request("get", "https://example.com", params=processed_params)
-    assert request.prepare().url.endswith(url_params)
-
-    # TODO: remove in 2.0.0
-    with pytest.deprecated_call():
-        processed_params = to_params_dict(option, value)
-
     request = requests.Request("get", "https://example.com", params=processed_params)
     assert request.prepare().url.endswith(url_params)
 
